flowedit: Move impl WindowState blocks to window_state.rs (#2611)
Move both impl WindowState blocks from flow_canvas.rs to
window_state.rs: the canvas message handler (handle_canvas_message
and extracted helpers) and the view builder (view_canvas_area and
overlay helpers). Clean up imports in both files.
